add extension method for DI for op data notifications
