+++ src/TokenCleanupHost.cs
@@ -1,0 +1,42 @@
+﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
+
+
+using IdentityServer4.EntityFramework.Options;
+using IdentityServer4.EntityFramework;
+using Microsoft.Extensions.Hosting;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.Extensions.DependencyInjection
+{
+    internal class TokenCleanupHost : IHostedService
+    {
+        private readonly TokenCleanup _tokenCleanup;
+        private readonly OperationalStoreOptions _options;
+
+        public TokenCleanupHost(TokenCleanup tokenCleanup, OperationalStoreOptions options)
+        {
+            _tokenCleanup = tokenCleanup;
+            _options = options;
+        }
+
+        public Task StartAsync(CancellationToken cancellationToken)
+        {
+            if (_options.EnableTokenCleanup)
+            {
+                _tokenCleanup.Start(cancellationToken);
+            }
+            return Task.CompletedTask;
+        }
+
+        public Task StopAsync(CancellationToken cancellationToken)
+        {
+            if (_options.EnableTokenCleanup)
+            {
+                _tokenCleanup.Stop();
+            }
+            return Task.CompletedTask;
+        }
+    }
+}
